+++ airsync-api/src/main/java/ffc/airsync/api/services/module/UserService.kt
@@ -30,7 +30,7 @@ object UserService {
             tokenMobile.insert(token = token,
               uuid = org.uuid,
               user = user,
-              id = id.toInt())
+              id = id)
             return TokenMessage(token)
         }
         throw NotAuthorizedException("Not Auth")
